feat: asset select refactor
