+++ src/lib/components/forms/assetInput/AssetInputNoOptionsMessage.tsx
@@ -2,7 +2,7 @@ import { Text } from "@chakra-ui/react";
 import type { NoticeProps } from "chakra-react-select";
 import { components } from "chakra-react-select";
 
-import type { AssetOption } from "./AssetInput";
+import type { AssetOption } from "lib/types";
 
 export const AssetInputNoOptionsMessage = (props: NoticeProps<AssetOption>) => {
   return (
